pyproject.toml
[project]
name = "iolaundry"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.5",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "scikit-learn>=1.7.1",
    "tensorboard>=2.20.0",
    "tensorboardx>=2.6.4",
    "torch-summary>=1.4.5",
    "tqdm>=4.67.1",
    "uvicorn>=0.35.0",
    "pyarrow>=21.0.0",
    "python-multipart>=0.0.20",
    "streamlit>=1.48.1",
    "streamlit-drawable-canvas>=0.9.3",
    "grad-cam>=1.5.5",
    "timm>=1.0.19",
]
